Divided into classes List and Array
